requirements.txt
flask
flask_sqlalchemy
python-dotenv
pyodbc
pymssql
requests
Pillow
google-generativeai